split by feature=system
