+++ Cargo.toml
@@ -4,17 +4,18 @@ version = "0.1.0"
 authors = ["Ogier Bouvier <ogier@bouvier.family>"]
 edition = "2018"
 license = "AGPL-3.0-or-later"
+resolver = "2"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-drop = { git = "https://github.com/Distributed-EPFL/drop", features = [ "system" ] }
+drop = { git = "https://github.com/Distributed-EPFL/drop" }
 sieve = { git = "https://github.com/Distributed-EPFL/sieve" }
 snafu = "0.6"
 tokio = { version = "1", features = [ "sync" ] }
 tokio-stream = "0.1"
 futures = "0.3"
-serde = "1.0"
+serde = "1"
 tracing = "0.1"
 tracing-futures = "0.2"
 derive_builder = "0.10"
@@ -29,3 +30,4 @@ tokio = { version = "1", features = [ "macros" ] }
 [features]
 default = []
 cli = [ "structopt", "sieve/cli" ]
+system = [ "drop/system", "sieve/system" ]
